requirements.txt
loguru==0.7.0
notion-client==2.0.0
requests==2.28.2
tenacity==8.2.2
python-dotenv==1.0.1
pytz==2024.2
PyGithub==2.5.0